# Conflicts:
#	src/SalesManager.java

+++ src/SalesManager.java
@@ -1,16 +1,16 @@
 import java.util.Arrays;
 
 public class SalesManager {
-  protected int[] sales;
+  protected long[] sales;
 
-  public SalesManager(int[] sales) {
+  public SalesManager(long[] sales) {
     Arrays.sort(sales);
     this.sales = sales;
   }
 
-  public int max() {
-    int max = -1;
-    for (int sale : sales) {
+  public long max() {
+    long max = -1;
+    for (long sale : sales) {
       if (sale > max) {
         max = sale;
       }
@@ -18,14 +18,14 @@ public class SalesManager {
     return max;
   }
 
-  public int truncatedMean() {
-    int mean = 0;
+  public long truncatedMean() {
+    long mean = 0;
 
     if(sales.length == 0) {
       return -1;
     }
 
-    for(int sale: sales) {
+    for(long sale: sales) {
       mean += sale;
     }
